Keep onnxruntime out of the app's web build
The preview CI builds the full app for web; pulling pdf_ocr_ondevice into
the app dragged onnxruntime (and ffi) into the dart2js compile, which fails
with "Only JS interop members may be 'external'" on their FFI bindings.

Guard the OCR import behind a conditional import (the same pattern the app
already uses for web_launch): ocr.dart now exports the native implementation
(ocr_native.dart, which imports pdf_ocr_ondevice) only where dart:io exists,
and a web stub (ocr_stub.dart, OnDeviceOcr.isSupported == false) otherwise —
so onnxruntime is never reached by the web build.

Verified: `flutter build web` for the app now succeeds; app tests pass.
